efuse: fix KASAN bug [1/1]
PD#TV-11119

Problem:
BUG: KASAN: stack-out-of-bounds in
meson64_trustzone_efuse+0x168/0x268

Solution:
fix retcnt variable type

Verify:
tl1 skt

+++ drivers/amlogic/efuse/efuse_hw64.c
@@ -185,7 +185,7 @@ ssize_t _efuse_read(char *buf, size_t count, loff_t *ppos)
 	unsigned int pos = *ppos;
 
 	struct efuse_hal_api_arg arg;
-	unsigned int retcnt;
+	unsigned long retcnt;
 	int ret;
 
 	arg.cmd = EFUSE_HAL_API_READ;
@@ -208,7 +208,7 @@ ssize_t _efuse_write(const char *buf, size_t count, loff_t *ppos)
 	unsigned int pos = *ppos;
 
 	struct efuse_hal_api_arg arg;
-	unsigned int retcnt;
+	unsigned long retcnt;
 	int ret;
 
 	arg.cmd = EFUSE_HAL_API_WRITE;
